rustc: New package
Add rustc-bootstrap, rust-std-bootstrap, cargo-bootstrap and the rustc package proper.

What I’m doing here is downloading the previous stable release of rust for our host machine and using that to bootstrap the latest release; after which we download the appropriate rust-std for our target; theoretically, we should be able to cross-compile rust programs using a host rustc and a target libstd.