assimp: Proper fix, just build irrxml ourselves
Decided to go with this because this thing has been a bug upstream for years in several different forms. Point is, I don’t think they can be trusted with IrrXML and since it’s a small library, it’s just cleaner to add it to mxe and declare it as a dependency for assimp.